# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyFlare API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  email: string;
  name: string;
  /** @nullable */
  avatar?: string | null;
}

export type UpdateProfileRequestPreferredGroupSize =
  (typeof UpdateProfileRequestPreferredGroupSize)[keyof typeof UpdateProfileRequestPreferredGroupSize];

export const UpdateProfileRequestPreferredGroupSize = {
  one_on_one: "one_on_one",
  small_group: "small_group",
  large_group: "large_group",
} as const;

export interface UpdateProfileRequest {
  name?: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  occupation?: string | null;
  /** @nullable */
  about?: string | null;
  /** @nullable */
  avatar?: string | null;
  interests?: string[];
  skills?: string[];
  preferredGroupSize?: UpdateProfileRequestPreferredGroupSize;
}

export interface UserProfile {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  occupation?: string | null;
  /** @nullable */
  about?: string | null;
  /** @nullable */
  avatar?: string | null;
  interests: string[];
  skills: string[];
  preferredGroupSize: string;
  totalSessions: number;
  totalHours: number;
  rating: number;
  createdAt: string;
}

export type BroadcastRequestGroupSize =
  (typeof BroadcastRequestGroupSize)[keyof typeof BroadcastRequestGroupSize];

export const BroadcastRequestGroupSize = {
  two_to_five: "two_to_five",
  five_to_ten: "five_to_ten",
} as const;

export interface BroadcastRequest {
  subject: string;
  latitude: number;
  longitude: number;
  groupSize: BroadcastRequestGroupSize;
  durationMinutes: number;
}

export interface NearbyStudent {
  id: number;
  name: string;
  /** @nullable */
  avatar?: string | null;
  subject: string;
  distanceKm: number;
  groupSize: string;
  interests: string[];
  rating: number;
  latitude: number;
  longitude: number;
}

export interface MatchResults {
  broadcastId: number;
  matches: NearbyStudent[];
}

export type StudySessionStatus =
  (typeof StudySessionStatus)[keyof typeof StudySessionStatus];

export const StudySessionStatus = {
  pending: "pending",
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface StudySession {
  id: number;
  subject: string;
  hostUserId: number;
  /** @nullable */
  partnerUserId?: number | null;
  status: StudySessionStatus;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  endedAt?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  createdAt: string;
}

export interface CreateSessionRequest {
  subject: string;
  /** @nullable */
  partnerUserId?: number | null;
  plannedDurationMinutes: number;
}

export type UpdateSessionRequestStatus =
  (typeof UpdateSessionRequestStatus)[keyof typeof UpdateSessionRequestStatus];

export const UpdateSessionRequestStatus = {
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateSessionRequest {
  status?: UpdateSessionRequestStatus;
  /** @nullable */
  durationMinutes?: number | null;
}

export interface FeedbackRequest {
  rating: number;
  /** @nullable */
  comment?: string | null;
  reviewedUserId: number;
}

export interface Feedback {
  id: number;
  sessionId: number;
  reviewerUserId: number;
  reviewedUserId: number;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface StudyGroup {
  id: number;
  name: string;
  subject: string;
  /** @nullable */
  description?: string | null;
  maxMembers: number;
  currentMembers: number;
  createdByUserId: number;
  createdAt: string;
}

export interface CreateGroupRequest {
  name: string;
  subject: string;
  /** @nullable */
  description?: string | null;
  maxMembers: number;
}

export interface StatsSummary {
  totalSessions: number;
  totalHoursStudied: number;
  averageRating: number;
  /** @nullable */
  topSubject?: string | null;
  completedThisWeek: number;
  activeGroups: number;
}

export type GetNearbyStudentsParams = {
  subject?: string;
  minMembers?: number;
  maxMembers?: number;
};

export type ListGroupsParams = {
  subject?: string;
};