# lib/api-zod/src/generated/types/broadcastRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyFlare API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BroadcastRequestGroupSize } from "./broadcastRequestGroupSize";

export interface BroadcastRequest {
  subject: string;
  latitude: number;
  longitude: number;
  groupSize: BroadcastRequestGroupSize;
  durationMinutes: number;
}