# lib/api-zod/src/generated/types/feedback.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyFlare API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Feedback {
  id: number;
  sessionId: number;
  reviewerUserId: number;
  reviewedUserId: number;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}