# lib/api-zod/src/generated/types/feedbackRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyFlare API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FeedbackRequest {
  rating: number;
  /** @nullable */
  comment?: string | null;
  reviewedUserId: number;
}