# lib/api-zod/src/generated/types/matchResults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyFlare API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NearbyStudent } from "./nearbyStudent";

export interface MatchResults {
  broadcastId: number;
  matches: NearbyStudent[];
}