# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyFlare API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalSessions: number;
  totalHoursStudied: number;
  averageRating: number;
  /** @nullable */
  topSubject?: string | null;
  completedThisWeek: number;
  activeGroups: number;
}