# lib/api-zod/src/generated/types/studyGroup.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyFlare API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudyGroup {
  id: number;
  name: string;
  subject: string;
  /** @nullable */
  description?: string | null;
  maxMembers: number;
  currentMembers: number;
  createdByUserId: number;
  createdAt: string;
}