# lib/api-zod/src/generated/types/studySession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyFlare API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudySessionStatus } from "./studySessionStatus";

export interface StudySession {
  id: number;
  subject: string;
  hostUserId: number;
  /** @nullable */
  partnerUserId?: number | null;
  status: StudySessionStatus;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  endedAt?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  createdAt: string;
}