# lib/api-zod/src/generated/types/updateSessionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyFlare API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateSessionRequestStatus } from "./updateSessionRequestStatus";

export interface UpdateSessionRequest {
  status?: UpdateSessionRequestStatus;
  /** @nullable */
  durationMinutes?: number | null;
}