# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyFlare API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  occupation?: string | null;
  /** @nullable */
  about?: string | null;
  /** @nullable */
  avatar?: string | null;
  interests: string[];
  skills: string[];
  preferredGroupSize: string;
  totalSessions: number;
  totalHours: number;
  rating: number;
  createdAt: string;
}